+++ odoo/songs/install/logistics.py
@@ -386,7 +386,7 @@ def create_picking_types(ctx):
          'color': color_ali,
          'sequence': 5,
          'zone_code': '04',
-         'food_type': True,
+         'is_portable_printer': True,
          },
         {'xmlid': '__setup__.stock_picking_type_medoc',
          'name': 'Pick Médicaments',